requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
rsa==4.6
six==1.15.0
telebot==0.0.4
Telethon==1.18.2
toml==0.10.2
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
